+++ test/fixtures/multifile/unexported_access/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "./mypkg"
+
+func main() {
+    println(mypkg.helper())
+}