+++ test/fixtures/multifile/unexported_access/mypkg/pkg.go
@@ -1,0 +1,9 @@
+package mypkg
+
+func Exported() int {
+    return helper()
+}
+
+func helper() int {
+    return 42
+}